feat: add cahute_get_info()
This is in preparation for when the Cahute library
will become a dynamic library for common systems,
rather than a static library. In such a case, it is
possible that the utilities are compiled with a
different version of the library, and use it with
headers that are not from the exact same version as
the compiled library.

As such, the Cahute library provides a new function
named cahute_get_info() to get the versioning
information from the compiled library, and a common
internal header <version_data.h> for utilities to get
their own versioning information.